lessons/l5/module8/timescale-integrations/requirements.txt
psycopg2-binary
requests
pandas
matplotlib
seaborn
python-dotenv
schedule